+++ Cargo.toml
@@ -19,10 +19,6 @@ default = []
 # that it should activate any useful Nightly things accordingly.
 unstable = []
 
-# Generate tests which are random inputs and the outputs are calculated with
-# musl libc.
-musl-reference-tests = ['rand']
-
 # Used to prevent using any intrinsics or arch-specific code.
 force-soft-floats = []
 
@@ -30,13 +26,16 @@ force-soft-floats = []
 members = [
   "crates/compiler-builtins-smoke-test",
   "crates/libm-bench",
+  "crates/libm-test",
+]
+default-members = [
+  ".",
+  "crates/libm-test",
 ]
 
 [dev-dependencies]
 no-panic = "0.1.8"
 
-[build-dependencies]
-rand = { version = "0.6.5", optional = true }
 
 # This is needed for no-panic to correctly detect the lack of panics
 [profile.release]
